requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.1
gunicorn==21.2.0
sqlmodel==0.0.16
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
psycopg[binary]==3.2.12
python-multipart==0.0.20
alembic==1.13.1
python-dotenv==1.2.1

# OpenAI
openai==1.54.0

# Integrations
twilio==9.3.7
boto3==1.35.0
sendgrid==6.11.0
redis==5.2.0

# Documents / media
reportlab==4.1.0
fpdf2==2.7.9
websockets==15.0.1

# HTTP / tasks
httpx==0.27.0
celery==5.4.0
dramatiq[redis]==1.17.0

# Utilities
python-dateutil==2.9.0
pytz==2024.1
email-validator==2.3.0
